always post-process non-AKN html
